tests: split up `VaultOps.test.ts`
